requirements.txt
gradio
requests
wikipedia
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
langgraph
rank_bm25
python-dotenv
IPython
langchain-google-genai
langchain-tavily
langchain-chroma
huggingface_hub
arxiv
pymupdf
pgvector